fixed GMAP calculation by casting denimonator within root to double

+++ project-team05/src/main/java/edu/cmu/lti/oaqa/annotators/Evaluator.java
@@ -110,7 +110,7 @@ public class Evaluator extends JCasAnnotator_ImplBase {
     for (Double val : vals) {
       result *= (val + epsilon);
     }
-    return Math.pow(result, 1/vals.size());
+    return Math.pow(result, 1/(double)vals.size());
   }
 
   /**
